requirements.txt
attrs==21.4.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.5.18.1
charset-normalizer==2.0.12
idna==3.3
JsonForm==0.0.2
jsonschema==4.5.1
JsonSir==0.0.2
pyenchant==3.2.2
pyrsistent==0.18.1
Python-EasyConfig==0.1.7
python-resources==0.3
PyYAML==6.0
requests==2.27.1
Resource==0.2.1
six==1.16.0
soupsieve==2.3.2.post1
urllib3==1.26.9
